requirements.txt
pygame
langdetect
gtts
SpeechRecognition
opencv-python
mediapipe
face_recognition
blob
pytesseract
python-dotenv